requirements.txt
alabaster==0.7.12
appnope==0.1.0
atomicwrites==1.3.0
attrs==19.3.0
Babel==2.7.0
backcall==0.1.0
blis==0.2.4
boto3==1.10.19
botocore==1.13.19
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
conllu==1.3.1
cycler==0.10.0
cymem==2.0.3
decorator==4.4.1
dialogflow==0.7.2
docutils==0.15.2
flaky==3.6.1
Flask==1.1.1
Flask-Cors==3.0.8
ftfy==5.6
gevent==1.4.0
google-api-core==1.14.3
google-auth==1.6.3
googleapis-common-protos==1.6.0
greenlet==0.4.15
grpcio==1.24.3
h5py==2.10.0
idna==2.8
imagesize==1.1.0
importlib-metadata==0.23
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.3
jmespath==0.9.4
joblib==0.14.0
jsonpickle==1.2
jsonschema==2.6.0
jupyter-client==5.3.4
jupyter-core==4.6.1
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
more-itertools==7.2.0
murmurhash==1.0.2
neuralcoref==4.0
nltk==3.4.5
numpy==1.17.4
numpydoc==0.9.1
overrides==2.5
packaging==19.2
pandas==0.25.3
parsimonious==0.8.1
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
plac==0.9.6
pluggy==0.13.0
preshed==2.0.1
prompt-toolkit==2.0.10
protobuf==3.10.0
psutil==5.6.5
ptyprocess==0.6.0
py==1.8.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
Pygments==2.4.2
pyparsing==2.4.5
pytest==5.2.4
python-dateutil==2.8.0
pytorch-pretrained-bert==0.6.2
pytorch-transformers==1.1.0
pytz==2019.3
pyzmq==18.1.1
readability==0.3.1
regex==2019.11.1
requests==2.22.0
responses==0.10.6
rsa==4.0
s3transfer==0.2.1
scikit-learn==0.20.3
scipy==1.3.2
sentencepiece==0.1.83
six==1.12.0
snowballstemmer==2.0.0
spacy==2.1.0
Sphinx==2.2.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sqlparse==0.3.0
srsly==0.2.0
stanfordcorenlp==3.9.1.1
tensorboardX==1.9
textblob==0.15.3
thinc==7.0.8
torch==1.3.1
tornado==6.0.3
tqdm==4.38.0
traitlets==4.3.3
Unidecode==1.1.1
urllib3==1.25.6
wasabi==0.4.0
wcwidth==0.1.7
Werkzeug==0.16.0
word2number==1.1
zipp==0.6.0
